Move customer reducer into its own module
